src/requirements.txt
certifi==2019.11.28
chardet==3.0.4
crypto==1.4.1
ecdsa==0.9
et-xmlfile==1.0.1
idna==2.8
jdcal==1.4.1
Naked==0.1.31
numpy==1.15.4
openpyxl==2.3.5
pandas==0.24.0
paramiko==1.12.0
pycrypto==2.6
python-dateutil==2.8.0
pytz==2018.9
PyVCF==0.6.8
PyYAML==3.11
requests==2.22.0
shellescape==3.4.1
six==1.12.0
suds==0.4
urllib3==1.25.8
